Move loader and web archive comps to ui folder
