+++ site/gatsby-site/src/components/TaxonomyForm.js
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import { useUserContext } from 'contexts/userContext';
 import { useMongo } from 'hooks/useMongo';
 import { Formik } from 'formik';
-import Loader from 'components/Loader';
+import Loader from 'components/ui/Loader';
 import config from '../../config.js';
 import { useMutation, useQuery } from '@apollo/client';
 import {
